CloudAccount: remove

+++ ui_backend/usecases/CloudAccount.py
@@ -10,15 +10,15 @@ class CloudAccount(Workflow):
     def __init__(self, username: str, workflowId: str, workflowAction: str, data: dict = None, headers: dict = None, *args, **kwargs):
         super().__init__(username, workflowId, *args, **kwargs)
 
-        self.workflowName = "cloud_account"
+        self.workflowName = "cloud_account-" + workflowAction
         self.username = username
         self.workflowId = workflowId
         self.workflowAction = workflowAction
         self.data = self.__dataformat(data)
         self.headers = headers or ()
 
-        self.calls = {
-            "assign": {
+        if workflowAction == "assign":
+            self.calls = {
                 "infobloxUnlock": {
                     "technology": "infoblox",
                     "method": "DELETE",
@@ -28,7 +28,7 @@ class CloudAccount(Workflow):
                 "checkpointDatacenterAccountPut" : {
                     "technology": "checkpoint",
                     "method": "PUT",
-                    "urlSegment": str(data.get("checkpoint_datacenter_account_put", {}).get("asset", 0)) + "/datacenter-account/",
+                    "urlSegment": str(data.get("checkpoint_datacenter_account_put", {}).get("asset", 0)) + "/datacenter-accounts/",
                     "data": self.data.get("checkpoint_datacenter_account_put", {})
                 },
                 "checkpointUnlock": {
@@ -38,17 +38,62 @@ class CloudAccount(Workflow):
                     "data": None
                 },
             }
-        }
-        # Each added infoblox network is a different call.
-        n = 0
-        for data in self.data.get("infoblox_assign_cloud_network", []):
-            self.calls["assign"]["infobloxAssignCloudNetwork-"+str(n)] = {
-                "technology": "infoblox",
-                "method": "PUT",
-                "urlSegment": str(data.get("asset", 0)) + "/assign-cloud-network/",
-                "data": data
+            # Each added infoblox network is a different call.
+            n = 0
+            for infobloxNetworkData in self.data.get("infoblox_cloud_network_assign", []):
+                self.calls["infobloxAssignCloudNetwork-" + str(n)] = {
+                    "technology": "infoblox",
+                    "method": "PUT",
+                    "urlSegment": str(infobloxNetworkData.get("asset", 0)) + "/assign-cloud-network/",
+                    "data": infobloxNetworkData
+                }
+                n += 1
+
+        elif workflowAction == "remove":
+            self.calls = {
+                "infobloxUnlock": {
+                    "technology": "infoblox",
+                    "method": "DELETE",
+                    "urlSegment": "locks/",
+                    "data": None
+                },
+                "checkpointDatacenterAccountDelete": {
+                    "technology": "checkpoint",
+                    "method": "DELETE",
+                    "urlSegment": str(data.get("checkpoint_datacenter_account_delete", {}).get("asset", 0)) + "/datacenter-account/" + str(data.get("Account Name", "")) + "/",
+                    "data": self.data.get("checkpoint_datacenter_account_delete", {}),
+                },
+                "checkpointUnlock": {
+                    "technology": "checkpoint",
+                    "method": "DELETE",
+                    "urlSegment": "locks/",
+                    "data": None
+                }
             }
-            n += 1
+            # Each added infoblox network is a different call.
+            i = 0
+            assetIds = []
+            for infobloxNetworkData in self.data.get("infoblox_cloud_network_delete", []):
+                assetIds.append(infobloxNetworkData.get("asset", 0))
+                self.calls["infobloxDeleteCloudNetwork-" + str(i)] = {
+                    "technology": "infoblox",
+                    "method": "DELETE",
+                    "urlSegment": str(infobloxNetworkData.get("asset", 0)) + "/delete-cloud-network/" + str(infobloxNetworkData.get("network", "")) + "/",
+                    "data": infobloxNetworkData
+                }
+                i += 1
+
+            assetIds = list(set(assetIds))
+            for id in assetIds:
+                if id:
+                    # Get info about the infoblox networks of the account on this asset.
+                    self.calls["infobloxAccountNetworksGet-" + str(id)] = {
+                        "technology": "infoblox",
+                        "method": "GET",
+                        "urlSegment": str(id) + "/networks/?fby=*Account Name&fval=" + self.data.get("Account Name", "") + "&fby=*Environment&fval=Cloud",
+                        "data": None
+                    }
+
 
     """
     "data": {                                             
@@ -57,7 +102,7 @@ class CloudAccount(Workflow):
         "Account ID": "555555555555",                     
         "provider": "AWS",  
         "Reference": "tizio",            
-        "infoblox_assign_cloud_network": [
+        "infoblox_cloud_network_assign": [
             {
                 "asset": 1,
                 "comment": "Nella vecchia fattoria ia ia oh",
@@ -87,8 +132,10 @@ class CloudAccount(Workflow):
 
     def preCheckPermissions(self) -> bool:
         try:
+            Log.log(self.data, 'DDDDDDDDDDDDDD')
+            Log.log(self.calls, 'CCCCCCCCCCCCCCCCC')
             self.checkAuthorizations()
-            self.checkWorkflowPrivileges(calls=self.calls[self.workflowAction])
+            self.checkWorkflowPrivileges(calls=self.calls)
 
             return True
         except Exception as e:
@@ -98,14 +145,13 @@ class CloudAccount(Workflow):
 
     def run(self) -> dict:
         response = None
-        calls = self.calls[self.workflowAction]
 
         try:
             if self.workflowAction == "assign":
-                for k in calls.keys():
+                for k in self.calls.keys():
                     if k.startswith("infobloxAssignCloudNetwork"):
                         response, status = self.requestFacade(
-                            **calls[k],
+                            **self.calls[k],
                             headers=self.headers,
                         )
                         Log.log("[WORKFLOW] " + self.workflowId + " - Infoblox response status: " + str(status))
@@ -115,10 +161,54 @@ class CloudAccount(Workflow):
                             raise CustomException(status=status, payload={"Infoblox": response})
                         else:
                             # Add the region in checkpoint data.
-                            calls["checkpointDatacenterAccountPut"]["data"]["regions"].append( calls[k]["data"]["region"].lstrip(calls[k]["data"]["provider"].lower() + '-') )
+                            self.calls["checkpointDatacenterAccountPut"]["data"]["regions"].append( self.calls[k]["data"]["region"].lstrip(self.calls[k]["data"]["provider"].lower() + '-') )
 
                 response, status = self.requestFacade(
-                    **calls["checkpointDatacenterAccountPut"],
+                    **self.calls["checkpointDatacenterAccountPut"],
+                    headers=self.headers,
+                )
+                Log.log("[WORKFLOW] " + self.workflowId + " - Checkpoint response status: " + str(status))
+                Log.log("[WORKFLOW] " + self.workflowId + " - Checkpoint response: " + str(response))
+
+                if status != 200:
+                    raise CustomException(status=status, payload={"Checkpoint": response})
+
+            elif self.workflowAction == "remove":
+                # Remove the infoblox networks.
+                m = max( [ int(k.lstrip('infobloxAccountNetworksGet-')) for k in self.calls.keys() if k.startswith("infobloxAccountNetworksGet") ] )
+                i = 0
+                while i <= m:
+                    response, status = self.requestFacade(
+                        **self.calls["infobloxDeleteCloudNetwork-" + str(i)],
+                        headers=self.headers,
+                    )
+                    Log.log("[WORKFLOW] " + self.workflowId + " - Infoblox response status: " + str(status))
+                    Log.log("[WORKFLOW] " + self.workflowId + " - Infoblox response: " + str(response))
+
+                    if status != 200:
+                        raise CustomException(status=status, payload={"Checkpoint": response})
+
+                # Now list the remaining regions.
+                regions = list()
+                for k in self.calls.keys():
+                    if k.startswith("infobloxAccountNetworksGet"):
+                        response, status = self.requestFacade(
+                            **self.calls[k],
+                            headers=self.headers,
+                        )
+                        Log.log("[WORKFLOW] " + self.workflowId + " - Infoblox response status: " + str(status))
+                        Log.log("[WORKFLOW] " + self.workflowId + " - Infoblox response: " + str(response))
+
+                        if status != 200:
+                            raise CustomException(status=status, payload={"Checkpoint": response})
+                        else:
+                            for net in response:
+                                regions.append( net.get("extattrs", {}).get("City", {}).get("value", "").lstrip( self.data.get("provider", "").lower() + "-") )
+
+                self.data["checkpoint_datacenter_account_delete"]["regions"] =  regions
+
+                response, status = self.requestFacade(
+                    **self.calls["checkpointDatacenterAccountDelete"],
                     headers=self.headers,
                 )
                 Log.log("[WORKFLOW] " + self.workflowId + " - Checkpoint response status: " + str(status))
@@ -133,7 +223,7 @@ class CloudAccount(Workflow):
         finally:
             # Release locks.
             r, s = self.requestFacade(
-                **calls["infobloxUnlock"],
+                **self.calls["infobloxUnlock"],
                 headers=self.headers,
             )
             if s == 200:
@@ -142,7 +232,7 @@ class CloudAccount(Workflow):
                 Log.log("Unlock failed on infoblox api: " + str(r))
 
             r, s = self.requestFacade(
-                **calls["checkpointUnlock"],
+                **self.calls["checkpointUnlock"],
                 headers=self.headers,
             )
             if s == 200:
@@ -157,12 +247,13 @@ class CloudAccount(Workflow):
     ####################################################################################################################
 
     def __dataformat(self, data: dict):
-        formattedData = {"infoblox_assign_cloud_network": [], "checkpoint_datacenter_account_put": {}}
+        formattedData = dict()
 
         try:
             if data:
                 if self.workflowAction == "assign":
-                    for network in data.get("infoblox_assign_cloud_network", []):
+                    formattedData = {"infoblox_cloud_network_assign": [], "checkpoint_datacenter_account_put": {}}
+                    for network in data.get("infoblox_cloud_network_assign", []):
                         dataItem = {"network_data": {}}
                         dataItem["asset"] = network.get("asset", 0)
                         dataItem["region"] = data.get("provider", "").lower() + "-" + network.get("region", "")
@@ -173,13 +264,22 @@ class CloudAccount(Workflow):
                         dataItem["network_data"]["extattrs"] = { "Reference": { "value": data.get("Reference", "")} }
                         dataItem["network_data"]["extattrs"]["Account Name"] = { "value": data.get("Account Name", "") }
                         dataItem["network_data"]["extattrs"]["Account ID"] = { "value": data.get("Account ID", "") }
-                        formattedData["infoblox_assign_cloud_network"].append(dataItem)
+                        formattedData["infoblox_cloud_network_assign"].append(dataItem)
 
                     formattedData["checkpoint_datacenter_account_put"] = {"change-request-id": data.get("change-request-id", "")}
                     formattedData["checkpoint_datacenter_account_put"]["Account Name"] = data.get("Account Name", "")
                     formattedData["checkpoint_datacenter_account_put"]["Account ID"] = data.get("Account ID", "")
                     formattedData["checkpoint_datacenter_account_put"]["tags"] = data.get("checkpoint_datacenter_account_put", {}).get("tags", [])
                     formattedData["checkpoint_datacenter_account_put"]["regions"] = [] # add each region in checkpoint data after the corresponding network is created in infoblox.
+                elif self.workflowAction == "remove":
+                    formattedData = {
+                        "Account Name": data.get("Account Name", ""),
+                        "provider": data.get("provider", ""),
+                        "infoblox_cloud_network_delete": data.get("infoblox_cloud_network_delete", []),
+                        "checkpoint_datacenter_account_delete": {
+                            "change-request-id": data.get("change-request-id", "")
+                        }
+                    }
 
             return formattedData
         except Exception as e:
